refactor(discover-v2): move DiscoverHome to feature module
Pure path migration — isolated so reviewers see only the move.
