+++ src/components/Discover/DiscoverScreenContent.tsx
@@ -4,9 +4,9 @@ import { View } from 'react-native';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
 
 import { useDiscoverSearchQueryStore } from '@/__swaps__/screens/Swap/resources/search/searchV2';
-import DiscoverHome from '@/components/Discover/DiscoverHome';
 import DiscoverSearch from '@/components/Discover/DiscoverSearch';
 import { Box, Inset } from '@/design-system';
+import DiscoverHome from '@/features/discover/components/DiscoverHome';
 
 export function DiscoverScreenContent() {
   const insets = useSafeAreaInsets();
